db create, and get all, extend reaction forwarding feature

+++ src/Entities.py
@@ -2,7 +2,7 @@ import uuid
 from datetime import datetime
 
 from pgvector.sqlalchemy import Vector
-from sqlalchemy import Integer
+from sqlalchemy import BigInteger
 from sqlalchemy.orm import DeclarativeBase
 from sqlalchemy.orm import mapped_column, Mapped
 
@@ -23,7 +23,7 @@ class BaseEntity(DeclarativeBase):
 class Message(BaseEntity):
     __tablename__ = "message"
 
-    member_id: Mapped[int] = mapped_column(Integer, nullable=False)
-    channel_id: Mapped[int] = mapped_column(Integer, nullable=False)
-    msg_id: Mapped[int] = mapped_column(Integer, nullable=False)
+    member_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
+    channel_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
+    msg_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
     embedding = mapped_column(Vector(1536), nullable=True)
